+++ src/index.js
@@ -9,8 +9,10 @@
  * OF ANY KIND, either express or implied. See the License for the specific language
  * governing permissions and limitations under the License.
  */
-const { openWhiskWrapper } = require('epsagon');
-const { wrap } = require('@adobe/helix-status');
+const { wrap: status } = require('@adobe/helix-status');
+const { wrap } = require('@adobe/openwhisk-action-utils');
+const { logger } = require('@adobe/openwhisk-action-logger');
+const { epsagon } = require('@adobe/helix-epsagon');
 const { execute } = require('./sendquery.js');
 const { cleanRequestParams } = require('./util.js');
 
@@ -44,35 +46,29 @@ async function runExec(params) {
   }
 }
 
-/**
- * Runs the action by wrapping the `runExec` function with the pingdom-status utility.
- * Additionally, if a EPSAGON_TOKEN is configured, the epsagon tracers are instrumented.
- * @param params Action params
- * @returns {Promise<*>} The response
- */
 async function run(params) {
-  let action = runExec;
-  action = openWhiskWrapper(action, {
-    token_param: 'EPSAGON_TOKEN',
-    appName: 'Helix Services',
-    metadataOnly: false, // Optional, send more trace data
-    ignoredKeys: ['token', /[A-Z0-9_]+/],
-  });
-  return wrap(action, {
-    fastly: 'https://api.fastly.com/public-ip-list',
-    googleiam: 'https://iam.googleapis.com/$discovery/rest?version=v1',
-    googlebigquery: 'https://www.googleapis.com/discovery/v1/apis/bigquery/v2/rest',
-  })(params);
-}
-
-async function main(params) {
-  if (params.__ow_headers && ('x-token' in params.__ow_headers) && ('x-service' in params.__ow_headers)) {
+  if (params.__ow_headers
+    && ('x-token' in params.__ow_headers)
+    && ('x-service' in params.__ow_headers)) {
     // eslint-disable-next-line no-param-reassign
     params.token = params.__ow_headers['x-token'];
     // eslint-disable-next-line no-param-reassign
     params.service = params.__ow_headers['x-service'];
   }
-  return run(params);
+  return runExec(params);
 }
 
-module.exports.main = main;
+/**
+ * Main function called by the openwhisk invoker.
+ * @param params Action params
+ * @returns {Promise<*>} The response
+ */
+module.exports.main = wrap(run)
+  .with(epsagon)
+  .with(status, {
+    fastly: 'https://api.fastly.com/public-ip-list',
+    googleiam: 'https://iam.googleapis.com/$discovery/rest?version=v1',
+    googlebigquery: 'https://www.googleapis.com/discovery/v1/apis/bigquery/v2/rest',
+  })
+  .with(logger.trace)
+  .with(logger);